stateless-chat-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.55.3
httpx==0.27.2
pydantic==2.6.0
python-dotenv==1.0.1
redis==5.0.1
